+++ CHANGELOG.md
@@ -9,11 +9,28 @@
 ### Added
 
 - `Style::rounding` for the rounding of the dock area border.
+- `TabStyle::active` for the active style of a tab.
+- `TabStyle::inactive` for the active style of a tab.
+- `TabStyle::focused` for the active style of a tab.
+- `TabStyle::hovered` for the active style of a tab.
 - `TabStyle::tab_body` for styling the body of the tab including background color, stroke color, rounding and inner margin.
+- `TabStyle::prefered_width` to set the prefered width of the tab.
+- `TabInteractionStyle` to style the active/inactive/focused/hovered states of a tab.
+
 
 ### Breaking changes
 
 - Moved `TabStyle::inner_margin` to `TabBodyStyle::inner_margin`.
+- Moved `TabStyle::fill_tab_bar` to `TabBarStyle::fill_tab_bar`.
+- Moved `TabStyle::outline_color` to `TabInteractionStyle::outline_color`.
+- Moved `TabStyle::rounding` to `TabInteractionStyle::rounding`.
+- Moved `TabStyle::bg_fill` to `TabInteractionStyle::bg_fill`.
+- Moved `TabStyle::text_color_unfocused` to `TabStyle::inactive.text_color`.
+- Moved `TabStyle::text_color_active_focused` to `TabStyle::focused.text_color`.
+- Moved `TabStyle::text_color_active_unfocused` to `TabStyle::active.text_color`.
+- Renamed `Style::tabs` to `Style::tab`
+- Removed `TabStyle::text_color_focused`. This style was pratically never reachable.
+
 
 ## 0.6.1 - 2023-05-29